slack-bot/requirements.txt
slack-bolt>=1.18.0
slack-sdk>=3.21.0
anthropic>=0.18.0
requests>=2.31.0
python-dotenv>=1.0.0
duckduckgo-search>=4.0.0
